# app/static/swagger.json
{
    "components" : {
        "securitySchemes" : {
          "BearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"}
          },
...
        },
    "servers" : [ {
      "url" : "https://viccnick.azurewebsites.net"
    } ],
    "openapi" : "3.0.1",
    "paths" : {
...
      "/api/users/{username}" : {
        "get" : {
          "description" : "Get specific user by username",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses" : {
            "200" : {
              "description" : "success",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "_links" : {
                        "type" : "object",
                        "properties" : {
                          "self" : {
                            "type" : "string"
                          }
                        }
                      },
                      "certificates_count" : {
                        "type" : "integer"
                      },
                      "id" : {
                        "type" : "integer"
                      },
                      "email" : {
                        "type" : "string"
                      },
                      "username" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"_links\":{\"self\":\"/api/users/5\"},\"certificates_count\":51,\"email\":\"testdata@test.com\",\"id\":5,\"username\":\"testdata\"}\n"
                    }
                  }
                }
              }
            },
            "401" : {
                "description" : "Unauthorized",
                "content" : {
                "application/json" : {
                    "schema" : {
                    "type" : "object",
                    "properties" : {
                        "error" : {
                        "type" : "string"
                        }
                    }
                    },
                    "examples" : {
                    "0" : {
                        "value" : "{\"error\":\"Unauthorized\"}\n"
                    }
                    }
                }
                }
            },
            "404" : {
                "description" : "Not Found",
                "content" : {
                "application/json" : {
                    "schema" : {
                    "type" : "object",
                    "properties" : {
                        "error" : {
                        "type" : "string"
                        }
                    }
                    },
                    "examples" : {
                    "0" : {
                        "value" : "{\"error\":\"Not Found\"}\n"
                    }
                    }
                }
                }
            },
            "500" : {
                "description" : "Internal Server Error",
                "content" : {
                "application/json" : {
                    "schema" : {
                    "type" : "object",
                    "properties" : {
                        "error" : {
                        "type" : "string"
                        }
                    }
                    },
                    "examples" : {
                    "0" : {
                        "value" : "{\"error\":\"Internal Server Error\"}\n"
                    }
                    }
                }
                }
            }
          },
          "parameters": [
            {
              "schema": {
                "nullable": false,
                "pattern": "^[a-zA-Z0-9_.]+$",
                "description": "username",
                "type": "string",
                "example": "testuser"
              },
              "in": "path",
              "name": "username",
              "description": "username",
              "required": true
            }
          ]
        },
        "put" : {
          "description" : "Edit own user",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "email" : {
                      "type" : "string"
                    },
                    "username" : {
                      "type" : "string"
                    }
                  }
                },
                "examples" : {
                  "0" : {
                    "value" : "{\n    \"username\": \"testdata\",\n    \"email\": \"testdata@test.com\"\n    \n}"
                  }
                }
              }
            }
          },
          "responses" : {
            "200" : {
              "description" : "success",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "_links" : {
                        "type" : "object",
                        "properties" : {
                          "self" : {
                            "type" : "string"
                          }
                        }
                      },
                      "certificates_count" : {
                        "type" : "integer"
                      },
                      "id" : {
                        "type" : "integer"
                      },
                      "email" : {
                        "type" : "string"
                      },
                      "username" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"_links\":{\"self\":\"/api/users/5\"},\"certificates_count\":51,\"email\":\"testdata@test.com\",\"id\":5,\"username\":\"testdata\"}\n"
                    }
                  }
                }
              }
            },
            "401" : {
                "description" : "Unauthorized",
                "content" : {
                "application/json" : {
                    "schema" : {
                    "type" : "object",
                    "properties" : {
                        "error" : {
                        "type" : "string"
                        }
                    }
                    },
                    "examples" : {
                    "0" : {
                        "value" : "{\"error\":\"Unauthorized\"}\n"
                    }
                    }
                }
                }
            },
            "404" : {
                "description" : "Not Found",
                "content" : {
                "application/json" : {
                    "schema" : {
                    "type" : "object",
                    "properties" : {
                        "error" : {
                        "type" : "string"
                        }
                    }
                    },
                    "examples" : {
                    "0" : {
                        "value" : "{\"error\":\"Not Found\"}\n"
                    }
                    }
                }
                }
            },
            "500" : {
                "description" : "Internal Server Error",
                "content" : {
                "application/json" : {
                    "schema" : {
                    "type" : "object",
                    "properties" : {
                        "error" : {
                        "type" : "string"
                        }
                    }
                    },
                    "examples" : {
                    "0" : {
                        "value" : "{\"error\":\"Internal Server Error\"}\n"
                    }
                    }
                }
                }
            }
          },
          "parameters": [
            {
              "schema": {
                "nullable": false,
                "pattern": "^[a-zA-Z0-9_.]+$",
                "description": "username",
                "type": "string",
                "example": "testuser"
              },
              "in": "path",
              "name": "username",
              "description": "username",
              "required": true
            }
          ]
        }
        },
...
      "/api/users" : {
        "get" : {
          "description" : "Get all users and number of certificates",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses" : {
            "200" : {
              "description" : "success",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "_links" : {
                        "type" : "object",
                        "properties" : {
                          "next" : { },
                          "prev" : { },
                          "self" : {
                            "type" : "string"
                          }
                        }
                      },
                      "_meta" : {
                        "type" : "object",
                        "properties" : {
                          "per_page" : {
                            "type" : "integer"
                          },
                          "page" : {
                            "type" : "integer"
                          },
                          "total_pages" : {
                            "type" : "integer"
                          },
                          "total_items" : {
                            "type" : "integer"
                          }
                        }
                      },
                      "items" : {
                        "type" : "array",
                        "items" : {
                          "type" : "object",
                          "properties" : {
                            "_links" : {
                              "type" : "object",
                              "properties" : {
                                "self" : {
                                  "type" : "string"
                                }
                              }
                            },
                            "certificates_count" : {
                              "type" : "integer"
                            },
                            "id" : {
                              "type" : "integer"
                            },
                            "email" : {
                              "type" : "string"
                            },
                            "username" : {
                              "type" : "string"
                            }
                          }
                        }
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"_links\":{\"next\":null,\"prev\":null,\"self\":\"/api/users?page=1&per_page=10\"},\"_meta\":{\"page\":1,\"per_page\":10,\"total_items\":7,\"total_pages\":1},\"items\":[{\"_links\":{\"self\":\"/api/users/2\"},\"certificates_count\":1,\"email\":\"admin@admin.ch\",\"id\":2,\"username\":\"admin\"},{\"_links\":{\"self\":\"/api/users/3\"},\"certificates_count\":0,\"email\":\"test2@test.com\",\"id\":3,\"username\":\"test2\"},{\"_links\":{\"self\":\"/api/users/4\"},\"certificates_count\":0,\"email\":\"apitestuser2@test.com\",\"id\":4,\"username\":\"apitestuser\"},{\"_links\":{\"self\":\"/api/users/5\"},\"certificates_count\":51,\"email\":\"testdata2@test.com\",\"id\":5,\"username\":\"testdata2\"},{\"_links\":{\"self\":\"/api/users/6\"},\"certificates_count\":0,\"email\":\"apicreateduser@test.com\",\"id\":6,\"username\":\"apicreateduser\"},{\"_links\":{\"self\":\"/api/users/7\"},\"certificates_count\":0,\"email\":\"testuser1@gmail.com\",\"id\":7,\"username\":\"testuser1\"},{\"_links\":{\"self\":\"/api/users/8\"},\"certificates_count\":0,\"email\":\"testuser12345@test.com\",\"id\":8,\"username\":\"testuser12345\"}]}\n"
                    }
                  }
                }
              }
            },
            "401" : {
              "description" : "Unauthorized",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "error" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"error\":\"Unauthorized\"}\n"
                    }
                  }
                }
              }
            },
            "404" : {
              "description" : "not found",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "error" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"error\":\"Not Found\"}\n"
                    }
                  }
                }
              }
            },
            "500" : {
              "description" : "Internal Server Error",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "error" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"error\":\"Internal Server Error\"}\n"
                    }
                  }
                }
              }
            }
          }
        },
        "post" : {
          "description" : "Create a User",
          "requestBody" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "password" : {
                      "type" : "string"
                    },
                    "email" : {
                      "type" : "string"
                    },
                    "username" : {
                      "type" : "string"
                    }
                  }
                },
                "examples" : {
                  "0" : {
                    "value" : "{\n    \"username\": \"testuser99999999\",\n    \"password\": \"password1\",\n    \"email\": \"testuserr000000@gmail.com\"\n    \n}"
                  },
                  "1" : {
                    "value" : "{\n    \"username\": \"testuser00000\",\n    \"password\": \"password1\",\n    \"email\": \"testuserr000000@gmail.com\"\n    \n}"
                  },
                  "2" : {
                    "value" : "{\n    \"username\": \"testuser00000\",\n    \"password\": \"password1\",\n    \"email\": \"testuserr000000@gmail.com\"\n    \n}"
                  }
                }
              }
            }
          },
          "responses" : {
            "201" : {
              "description" : "success",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "_links" : {
                        "type" : "object",
                        "properties" : {
                          "self" : {
                            "type" : "string"
                          }
                        }
                      },
                      "certificates_count" : {
                        "type" : "integer"
                      },
                      "id" : {
                        "type" : "integer"
                      },
                      "email" : {
                        "type" : "string"
                      },
                      "username" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"_links\":{\"self\":\"/api/users/9\"},\"certificates_count\":0,\"email\":\"testuserr000000@gmail.com\",\"id\":9,\"username\":\"testuser00000\"}\n"
                    }
                  }
                }
              }
            },
            "400" : {
              "description" : "bad request",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "error" : {
                        "type" : "string"
                      },
                      "message" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"error\":\"Bad Request\",\"message\":\"please use a different email address\"}\n"
                    },
                    "1" : {
                      "value" : "{\"error\":\"Bad Request\",\"message\":\"please use a different username\"}\n"
                    }
                  }
                }
              }
            },
            "404" :{
              "description" : "not found",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "error" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"error\":\"Not Found\"}\n"
                    }
                  }
                }
              }
            },
            "500" : {
              "description" : "Internal Server Error",
              "content" : {
                "application/json" : {
                  "schema" : {
                    "type" : "object",
                    "properties" : {
                      "error" : {
                        "type" : "string"
                      }
                    }
                  },
                  "examples" : {
                    "0" : {
                      "value" : "{\"error\":\"Internal Server Error\"}\n"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      }
    }